Fixed a compiler warning that was introduced in r19322 (patch 3044512). The third argument passed to read_objid() was an int instead of size_t. This could have caused memory corruption on 64-bit systems.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@19371 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/mibII/system_mib.c
@@ -168,7 +168,7 @@ system_parse_config_sysServices(const char *token, char *cptr)
 static void
 system_parse_config_sysObjectID(const char *token, char *cptr)
 {
-    int sysObjectIDLength = MAX_OID_LEN;
+    size_t sysObjectIDLength = MAX_OID_LEN;
     if (!read_objid(cptr, sysObjectID, &sysObjectIDLength)) {
 	netsnmp_config_error("sysobjectid token not a parsable OID:\n\t%s",
 			     cptr);
